feat: add copy_canceled signal to FileCopy and Worker classes
The `FileCopy` class now emits a `copy_canceled` signal when the copy process is canceled. This signal is connected to the `copy_canceled` slot in the `MainController` and `MainWindow` classes.

The `Worker` class now emits a `copy_canceled` signal when the copy process is canceled. This signal is connected to the `copy_canceled` slot in the `FileCopy` class.

This change allows the application to handle the case when the copy process is canceled by the user.